FCR summary of nedap pig performance station

FCR summary of nedap pig performance station

fcr_summary(data)

Arguments

data

A data.frame or a data.table. The data should include at least one pair of weight columns ("min_weight_cut" and "max_weight_cut" or "min_weight_origin" and "max_weight_origin") and the following common columns: "lm_slope", "test_days", "origin_dfi", "corrected_dfi", "fcr". The column "cor_fcr" is optional.

Value

A data.frame that includes the statistics of each column.

Examples

list_csv <- list.files("C:/Users/Dell/Downloads/location301", pattern = ".csv", full.names = TRUE)
data <- import_csv(list_csv)
fcr_res <- fcr_get(data = data, my_break = c(30, 100))
#> • There are no duplicate responders in different locations.
#> • The removing of weight < 15kg will not delete responder.
#> • Revoming records of missing will delete responders:  17 
#> • Deleted responders: 
#>  c("2310133","2310151","2310199","2310241","2310263","2310293","2310315","2310328","2310367","2310368","2310508","2310533","2310539","2310546","2310593","2310602","2310718") 
#> • Check for errors and warnings in RANSAC Robust
#>  - No errors encountered during processing.
#>  - No warnings encountered during processing.
#>  - RANSAC Robust succeeded
#> • The outliers detected by model will not delete responder.
#> • All responders' begin_test_weight are less than or equal to 60kg.
#> • All responders' end_test_weight are more than or equal to 85kg.
#> • Check for errors and warnings in Simple Linear Regression
#>  - No errors encountered during processing.
#>  - No warnings encountered during processing.
#> • Calculate ADG at 30~100kg weight range using Simple Linear Regression succeeded!
#> • Successfully generated the following 7 variables during the food intake correction process.
#>  - FIV:feed intake per visit;
#>  - OTV:occupation time per visit;
#>  - FRV:feeding rate per visit;
#>  - LWD:leading weight difference;
#>  - LTD:leading time difference;
#>  - FWD:following weight difference;
#>  - FTD:following time difference.
#> • Successfully generated 16 error types from 7 variables:
#>  - FIV-lo; FIV-hi; FIV-0; OTV-lo; OTV-hi; FRV-hi-FIV-lo; FRV-hi-strict; FRV-hi; FRV-0; FRV-lo;
#>  - LWD-lo; LWD-hi; FWD-lo; FWD-hi; LTD-lo; FTD-lo.
#> 
#> Warning: the standard deviation is zero
#> Warning: The correlation matrix has missing values. 27 columns were excluded from the
#> filter.
#> • Running model with equation: 
#> dfi_right_part ~ otd_10 + frv_0_p + frv_lo_p + location + lm_slope + weight + (1 | responder) 
#> • Running model with equation: 
#> dfi_right_part ~ otd_10 + frv_0_p + frv_lo_p + (1 | responder) 
#> Warning: number of items to replace is not a multiple of replacement length
fcr_summary(fcr_res[[1]])
#> # A tibble: 8 × 7
#>   set                N min     max     median  `mean+/-sd`      cv    
#>   <chr>          <dbl> <chr>   <chr>   <chr>   <chr>            <chr> 
#> 1 min_weight_cut    14 28.50   35.00   30.25   30.95+/-1.73     5.58% 
#> 2 max_weight_cut    14 91.50   102.50  99.50   98.80+/-3.44     3.49% 
#> 3 lm_slope          14 744.25  1271.62 1007.31 992.09+/-159.73  16.10%
#> 4 test_days         14 63.00   103.00  85.50   84.21+/-12.87    15.29%
#> 5 origin_dfi        14 1633.60 2873.46 2396.75 2336.10+/-360.62 15.44%
#> 6 corrected_dfi     14 1655.60 2856.33 2404.41 2333.63+/-345.74 14.82%
#> 7 fcr               14 2.14    2.55    2.36    2.36+/-0.15      6.43% 
#> 8 cor_fcr           14 2.14    2.53    2.38    2.36+/-0.15      6.18%